refactor: add composable to index
