+++ frontend/src/components/Layout/AudioControls.vue
@@ -95,7 +95,7 @@
 <script setup lang="ts">
 import { getItemDetailsLink } from '@/utils/items';
 import { playbackManagerStore } from '@/store';
-import { usePlayerKeys } from '@/composables/use-playerkeys';
+import { usePlayerKeys } from '@/composables';
 
 usePlayerKeys(false);
